Add logic for translating zone/zones/allowtopologies
