+++ staging/src/k8s.io/csi-translation-lib/plugins/BUILD
@@ -43,6 +43,7 @@ go_test(
         "azure_disk_test.go",
         "azure_file_test.go",
         "gce_pd_test.go",
+        "in_tree_volume_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
